+++ game_patch/misc/player.h
@@ -40,9 +40,17 @@ struct PlayerAdditionalData
     int last_activity_ms = 0;
     rf::TimestampRealtime idle_check_timestamp;
     rf::TimestampRealtime idle_kick_timestamp;
+    rf::Timestamp respawn_timer; // only used when configured in ADS
     uint8_t damage_handicap = 0; // percentile
 };
 
+static rf::Timestamp g_respawn_timer_local;
+static bool g_spawned_in_current_level = false; // relevant if force respawn is on
+static bool g_local_queued_delayed_spawn = false;
+
+std::string build_local_spawn_string(bool can_respawn);
+void set_local_spawn_delay(bool can_respawn, bool force_respawn, int spawn_delay);
+void reset_local_delayed_spawn();
 void find_player(const StringMatcher& query, std::function<void(rf::Player*)> consumer);
 void reset_player_additional_data(const rf::Player* player);
 PlayerAdditionalData& get_player_additional_data(rf::Player* player);
@@ -50,5 +58,6 @@ void play_local_hit_sound(bool died);
 void handle_chat_message_sound(std::string message);
 bool is_player_minimum_af_client_version(rf::Player* player, int version_major, int version_minor);
 bool is_server_minimum_af_version(int version_major, int version_minor);
+void player_multi_level_post_init();
 void update_player_flashlight();
 void ping_looked_at_location();